server/pyproject.toml
[project]
name = "mini-social-media"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.121.2",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "imagekitio>=4.2.0",
    "passlib[cryptography]>=1.7.4",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "uvicorn[standard]>=0.38.0",
]
